Using -Wno-missing-field-initializers instead of specifying a number of zeros, which is not portable for system structs. Fix more complaints due to -Wextra in debug mode. Enabled clean compiling in C++ mode with CC=g++-6 -fpremissive. Most of the errors were related to conversion from void* to other types. A few bugs were spotted and fixed.

+++ lib/turbulence.h
@@ -40,7 +40,7 @@ typedef struct GENATM_T{
     dmat *spect;     /**<The turbulence spectrum, sqrt of PSD*/
     long method;     /**<The method*/
 }GENATM_T;
-map_t *genatm_simple(double r0, double L0, double dx, double nx);
+map_t *genatm_simple(double r0, double L0, double dx, long nx);
 dmat *genatm_loc(loc_t *loc, double r0, double L0);
 mapcell* genatm_from_spect(GENATM_T *data);
 mapcell* vonkarman_screen(GENATM_T *data);
